pyproject.toml
[tool.poetry]
name = "chat_classifier"
version = "0.1.0"
description = ""
authors = ["anirudh-murali <anirudhvmurali@gmail.com>"]
readme = "README.md"
packages = [{include = "chat_classifier"}]

[tool.poetry.dependencies]
python = "3.10.0"
scipy = "^1.11.3"
tenacity = "^8.2.3"
tiktoken = "^0.5.1"
termcolor = "^2.3.0"
openai = "^0.28.1"
requests = "^2.31.0"
faiss-cpu = "^1.7.4"
matplotlib = "^3.8.0"
plotly = "^5.17.0"
plotly-express = "^0.4.1"
pandas = "^2.1.1"
scikit-learn = "^1.3.1"
sentence-transformers = "^2.2.2"


[tool.poetry.group.dev.dependencies]
ipython = "^8.16.1"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
